switch to ubuntu:16.04 for linux/android build
Since `ubuntu:xenial-20151218.1` uses a Schema 1 manifest,
which is no longer supported in newer versions of containerd,
we chose not to rely on `CONTAINERD_ENABLE_DEPRECATED_PULL_SCHEMA_1_IMAGE=1`,
as this option will be completely removed in containerd v2.1.

UID 102 -> 105 change needed as jenkins user on Debian 12 is running
under 102.